Two major fixes (SQL date query error and GPKG functionality) and minor changes (#23)
* Updated ogr script, executes the string in script

* updated ogr script, executes the string in script

* Updated script now works in all DTS desktops , no arcpy required

* script now works in all DTS desktops, no arcpy, corrects SQL dialect to work with ogr2ogr

* created readme.md for ogrFromBCGW

* updated ogrFromBCGW script, modified the readme.md
